Move cart items to redux

+++ client/src/components/Product.js
@@ -9,35 +9,21 @@ const Product = ({ product, onAddToCart }) => {
   const handleOnDeleteProduct = async (e) => {
     e.preventDefault();
 
-    // if (!window.confirm(`Are you sure you want to delete ${product.title} ?`)) {
-    //   return;
-    // }
+    if (!window.confirm(`Are you sure you want to delete ${product.title} ?`)) {
+      return;
+    }
 
     const productID = product._id;
     await axios.delete(`/api/products/${productID}`);
 
     dispatch(deleteProduct(productID));
   };
-  /*
-    - POST to server to delete product
-    - Trigger delete action
-    - In reducer, remove the product from the products
-
-  */
-  // const handleDeleteProduct = async (productId, callback) => {
-  //   await axios.delete(`/api/products/${productId}`);
-  //   /*
-  //   setProducts(products.filter((product) => product._id !== productId));
-  //   if (callback) {
-  //     callback();
-  //   }
-  //   */
-  // };
 
   const handleAddToCart = (e) => {
     e.preventDefault();
     onAddToCart(product._id);
   };
+
   return (
     <div className="product">
       <div className="product-details">
